requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy[asyncio]==2.0.42
asyncpg==0.30.0
pydantic==2.11.1
pydantic-settings==2.10.1
python-jose==3.5.0
passlib[bcrypt]==1.7.4
typing-extensions==4.14.1
alembic==1.16.4
